+++ apps/backend/src/modules/routes/route.service.ts
@@ -10,7 +10,7 @@ import {
 import { type CollectionResult, type Service } from "~/libs/types/types.js";
 
 import { type FileService } from "../files/files.service.js";
-import { type PlannedPathResponseDto } from "../planned-paths/libs/types/planned-path-response-dto.type.js";
+import { type PlannedPathResponseDto } from "../planned-paths/libs/types/types.js";
 import { type PlannedPathService } from "../planned-paths/planned-paths.js";
 import { type PointsOfInterestService } from "../points-of-interest/points-of-interest.service.js";
 import { RoutesExceptionMessage } from "./libs/enums/enums.js";