+++ apps/backend/src/modules/users/user.controller.test.ts
@@ -34,6 +34,7 @@ describe("UserController", () => {
 	it("findAll should return all users", async () => {
 		const users = [
 			{
+				avatarUrl: "https://aws/avatars/example_file.jpg",
 				email: "test@example.com",
 				firstName: "John",
 				group: mockGroup.toObject(),
@@ -60,6 +61,53 @@ describe("UserController", () => {
 			payload: {
 				data: users,
 			},
+			status: HTTPCode.OK,
+		});
+	});
+
+	it("getUserProfile should return user public profile", async () => {
+		const userProfile = {
+			avatarUrl: "https://aws/avatars/example_file.jpg",
+			firstName: "John",
+			followersCount: 5,
+			id: 1,
+			isFollowed: true,
+			isVisibleProfile: true,
+			lastName: "Doe",
+		};
+
+		const mockGetUserProfile: UserService["getUserProfile"] = () => {
+			return Promise.resolve(userProfile);
+		};
+
+		const userService = {
+			getUserProfile: mockGetUserProfile,
+		} as UserService;
+
+		const userController = new UserController(mockLogger, userService);
+
+		const mockUser = {
+			avatarUrl: null,
+			email: "test@example.com",
+			firstName: "Test",
+			group: mockGroup.toObject(),
+			groupId: 1,
+			id: 2,
+			isVisibleProfile: true,
+			lastName: "User",
+		};
+
+		const request = {
+			body: undefined,
+			params: { id: 1 },
+			query: undefined,
+			user: mockUser,
+		};
+
+		const result = await userController.getUserProfile(request);
+
+		assert.deepStrictEqual(result, {
+			payload: { data: userProfile },
 			status: HTTPCode.OK,
 		});
 	});