+++ apps/frontend/src/libs/modules/store/libs/types/root-reducer.type.ts
@@ -7,19 +7,23 @@ import {
 	type pointsOfInterestReducer,
 } from "~/modules/points-of-interest/points-of-interest.js";
 import {
+	type constructRouteReducer,
 	type routeDetailsReducer,
 	type routesReducer,
 } from "~/modules/routes/routes.js";
+import { type reducer as usersReducer } from "~/modules/users/users.js";
 
 type RootReducer = {
 	app: ReturnType<typeof appReducer>;
 	auth: ReturnType<typeof authReducer>;
+	constructRoute: ReturnType<typeof constructRouteReducer>;
 	explore: ReturnType<typeof exploreReducer>;
 	location: ReturnType<typeof locationReducer>;
 	pointOfInterestDetails: ReturnType<typeof pointOfInterestDetailsReducer>;
 	pointsOfInterest: ReturnType<typeof pointsOfInterestReducer>;
 	routeDetails: ReturnType<typeof routeDetailsReducer>;
 	routes: ReturnType<typeof routesReducer>;
+	users: ReturnType<typeof usersReducer>;
 };
 
 export { type RootReducer };