+++ packages/shared/src/modules/users/libs/enums/user-exception-message.enum.ts
@@ -2,6 +2,7 @@ const UserExceptionMessage = {
 	ALREADY_EXISTS: "A user with this email already exists",
 	INVALID_CREDENTIALS: "Invalid credentials",
 	NOT_FOUND: "User not found",
+	USER_ALLREADY_EXISTS: "A user with this email already exists",
 } as const;
 
 export { UserExceptionMessage };